Fix arrays of records.
The previous patch didn't handle arrays of rtecords, and core
dumped on some array inputs. We detect arrays of anonymous records
by checking for the RECORDARRYOID type, and for all arrays
get the input data for their base types. Arguably, this would be
better done for array types at plv8_type construction time. That
is lefts as a (hopefully very hear future) optimization.

The test case for this I have been using is:
      select q2json('select $$a$$ || x as b,
                            y as c,
                            array[row(x.*,array[1,2,3]),
                                  row(y.*,array[4,5,6])] as z
                     from generate_series(1,2) x,
                          generate_series(4,5) y');