Add the separate context mode.
The feature is currently experimental and users need set a custom
variable plv8.use_separate_context to on.  When off (as before
and default), one unique global context is stored in the heap
space and every function call uses it.  If a function modifies it,
the ramification is spread across other functions.  To avoid it,
when on, each function invocation creates its own context and run
inside it.  The downside of this feature is the overhead;  creating
a new Context is known to be expensive, and a benchmark shows it
takes much longer than a typical hash_search.  It may be ok in some
circomstances as the overhead is per-query-per-function.  In
typical SQL use cases, the cost is more in per-tuple call and this
feature doesn't add overhead for per-tuple calls.