New function plv8.find_function().
This function returns a javascript function that is registered in
the catalog as a plv8 function.  The key to find is a string of
either regproc or regprocedure.  This function invokes v8's function
creation, so not so fast.  So we'll need some caching mechanism
in either implicit (internal cache in plv8_proc) way or explicit
(user may cache it by using this object) way, in the latter of which
we'll need a function-local object in even the global-context mode.