Support scalar SRF.
Changes are around Converter and CreateTupleStore.  For scalar SRFs,
the backend creates a TupleDesc with a single attribute and pass it
to the function, so the procedure doesn't need to take care of
scalar case so much.  One note is ReturnSetInfo::expectedDesc is
not living for long, so we copy it to the per-query-context.  Also,
Add a check for non-object argument to record-type function's return.