Add user-separated context.
This enables to switch the global context based on the current user id,
which has been demanded from the security point of view.  A global
context in V8 is actually like an interpreter in other PLs so we only
needed to switch the global context.  We use std::vector instead of
postgres hash table for the user id search, because the size of array
should be relatively small and be more efficient in most cases.  Also,
move the recent startup facility to the global context initialization
process, as this now seems like the right place to run it, which leads
to extract find_js_function_by_name from plv8_FindFunction as this method
may be useful for more wide use.