Remove the separate context feature.
As we now have per-user context feature, the separate context approach
is not necessary anymore.  plv8_proc_cache now has function handle
instead of script handle, which was the original code until the
separate context.  Still, we need a per-function object namely "this"
and keep the exeucte environment as it is.  We may want to rethink this
structure (especially, transaction callback) when we come up with
better idea about how we provide per-function state storage, which
is "this" today.