Remove old style functions.
One note here is that plv8.return_next is now always a property
of the plv8 object even if the function is not an SRF.  In the
old style, yield/return_next function was created on demand,
but now only Converter and Tuplestore are created on SRF call.
The function itself is registered in the context initialization.