Let parser deduce parameter types in SPI queries.
The mechanism is based on the variable parameter hook available since 9.0.
For 8.4, we still have to specify the parameter types.  plan.execute,
plan.cursor is also updated to make use of this.

Change ValueGetDatum and InferredDatumType to lower_case functions, as
they look safe and natural to use outside of C++ try block.  plv8_param.cc
is also a bit modified to reduce code complexity in plv8_func.cc.  In 8.4,
plv8_param_state is never used, but leaving it in the source makes more
readable than hiding the definition completely.