Stricter checking for SRF returns.
We check that the right number of attributes are returned.

Still to do is checking that the names match one to one.
(The infrastructure is there, I just need to brush the cobwebs
off my C++ to write the actual name comparison test, and it's
too late at night for me to do that now.)