Remove dependency of v8/d8 to parse META.json.
Ideally we want to have only one place for the version string, but parsing
META.json is a bit challenging; at one time we had v8/d8 parsing META.json
to pass this value to source file, but it turned out those utilities may not
be available everywhere.  Since this integrity matters only developers,
we just check it if they are available.  We may come up with a better
solution to have it in one place in the future.