Support static link to v8.
In some circumstances, it is convenient to link v8 statically while we still
build plv8 with dynamic linking by default.  v8 supports both of dynamic and
static links, and also two modes of snapshot on and snapshot off in the static
library.  It might be possible to find the mode by looking at the v8 directory,
but it is wise to let users explicitly choose it to avoid to shoot in the foot.

Also rethink V8DIR, which was not working well.  Previous coding was not so
nice as the default value was quite unportable.  Now we have two different
variables, V8_SRCDIR pointing to the v8 source directory, and V8_OUTDIR
pointing to the v8 build output directory.  By default without specifying
them, the system will find the appropriate ones.  We might want to introduce
our own configure script, but for now we stay in Makefile which is still
simpler than other solutions.