Support json type conversion.
Now the json type is built-in since 9.2, it is reasonable to convert
the text representation to/from object.  V8 doesn't expose JsonParser
interface to C++ client, so we ended up extracting JSON.parse/stringify
from the global context.  There shouldn't be any performance gain
over the user script calling the functions, but it is only for convenience.
This breaks the existing script that calls JSON.parse/stringify.