Support v8 remote debugger.
The feature is enabled only if make is given ENABLE_DEBUGGER_SUPPORT=1,
because we are not sure how much of v8 installations are with debug
enabled.  Maybe it is safe to flip the flag to DISABLE_DEBUGGER_SUPPORT.
On the other hand, with static build, we are sure we can support it.