Introduce external array suport.
v8 has ever invented TypedArray, which maps JS array to fixed-length
element memory, originally for canvas pixel support.  We utilize it
for bytea conversion and int/float array.  In the array case, we had
to annotate the type by domain as it's an external array type to JS,
otherwise, we would loose the current flexibility around array conversion.
Those annotation type names may be changed in the future.  Also, it
might be reasonable to introduce TypedArray such like in d8, in order
for users to make their own fixed-length element array, but it will be
a different story.