Support window function API.
These functions help create user-defined window functions in PL/v8.
Most of the wrapper are straightforward, but partition local memory
and "isout" of get_func() are a little ugly.  For partition local
memory, we end up serializing the object by JSON.stringify, otherwise
v8's object lifecycle doesn't match to partition memory lifecycle.
"isout" is true if the value at the row is out of partition/frame,
but we didn't have a good way to return this.  Instead, because the
possible value to be returned in non-isout case is either value itself
or null, we return undefined in "isout" case.

In passing, fix a potential bug in SRF; if the function errors out
with exception, the restoration of the conv/tupstore will be omitted.
We make use of deconstructor not to leak them.