Update plv8.md
Added notes regarding problems I ran into on Ubuntu 20.04 LTS - I am not sure that I have placed them in the right part of the document.

This file is out of date and needs some love. The apt-get install postgresql-{your-postgresql-version-here}-plv8 instructions no longer seems to be valid.

Would it make sense to structure the document so that it had a section per operating system? The dependencies and instructions are very different in each OS.