Initialize plv8_context with zeroes. (#504)
Otherwise, the `unhandled_promises` vector can be initialized with garbage.
A better solution would be something like placement new, but there are
concerns about placement new support in old versions of MSVC. With zeroing
out during the initialization, we depend on the exact implementation of the
vector. As far as I know, zeroing out will set `len/cap` or `start/end` to
zero and will work correctly afterward.