Add workflow to release on PGXN for a semver tag (#570)
Just set `PGXN_USERNAME` and `PGXN_PASSWORD` in the project secrets and
it should just do it on the next semver tag.

Also use the `GIT_ARCHIVE_CMD` variable so that the bundle includes all
submodules, teach PGXN not to index them, and tweak the `Makefile` to
avoid calling `git` if the submodule is already included. Fixes #569.