Fix possible overflow in Window Functions (#591)
* Fix possible heap overflow in Window Functions

The issue is that the variable `size` is of the type `size_t`, which is an unsigned integer. 

Passing a negative value leads to an integer overflow, and the requested size is in the order of 2^64 bytes, which is over 10^10 GB, leading to a heap overflow. 

Calling the window function will now lead to an error (ERROR:  allocation size cannot be negative) instead of crashing.

* Add v8 namespace to String usage to fix build errors.